feat(#289) :: 트렌젝션 로직과 파일 업로드 로직 분류
